requirements.txt
fastapi
uvicorn
pydantic
aiohttp
pdfplumber
sentence-transformers
qdrant-client
openai
google-generativeai
numpy
tqdm
python-multipart
jinja2
python-dotenv
rank-bm25
transformers
torch

openpyxl
python-docx
python-pptx
pillow
aiohttp

dotenv

tabulate
pytesseract
opencv-python
pillow
langchain-groq
langchain-openai
langchain
langchain-google-genai
google-generativeai
bs4
aiodns

pymupdf
pandas